refactor: replace eps import from utils.math to math.machine for consistency
