+++ src/polykin/copolymerization/penultimate.py
@@ -9,7 +9,7 @@ from typing import Literal
 import numpy as np
 
 from polykin.kinetics import Arrhenius
-from polykin.utils.math import eps
+from polykin.math.machine import eps
 from polykin.utils.tools import check_bounds
 from polykin.utils.typing import FloatArray, FloatArrayLike, IntArrayLike
