+++ src/polykin/math/misc/jcr.py
@@ -12,9 +12,9 @@ from numpy import arctan, cos, exp, log, sin, sqrt
 from scipy.stats.distributions import f as Fdist
 from scipy.stats.distributions import t as tdist
 
+from polykin.math.machine import eps
 from polykin.math.roots import RootResult, root_secant
 from polykin.utils.exceptions import ShapeError
-from polykin.utils.math import eps
 from polykin.utils.tools import check_bounds
 from polykin.utils.typing import Float2x2Matrix, FloatVector
